Cleanup

+++ random.h
@@ -3,5 +3,5 @@
 #include <cstdlib>
 
 inline float rand_float(float min, float max) {
-	return (max - min) * static_cast<float>(rand()) / RAND_MAX + min;
+	return (max - min) * (static_cast<double>(rand()) / RAND_MAX) + min;
 }
